+++ semcore/mini-chart/__tests__/mini-chart.browser-test.tsx
@@ -1,84 +1,252 @@
 import { e2eStandToHtml } from '@semcore/testing-utils/e2e-stand';
 import { expect, test } from '@semcore/testing-utils/playwright';
 
-test.describe('MiniChart', () => {
-  test.describe('Score', () => {
-    test('render Lines', async ({ page }) => {
-      const standPath = 'stories/components/mini-chart/tests/examples/score-lines.tsx';
-      const htmlContent = await e2eStandToHtml(standPath, 'en');
+test.describe('Visual-Score Donuts', () => {
+  const ScoreDonut = [
+    { value: 0, color: undefined, baseBgColor: undefined },
+    { value: 0.5, color: 'chart-palette-order-2', baseBgColor: 'chart-palette-order-10' },
+    { value: 1, color: 'chart-palette-order-3', baseBgColor: 'chart-palette-order-11' },
+    { value: 33, color: 'chart-palette-order-4', baseBgColor: undefined },
+    { value: 69.5, color: undefined, baseBgColor: 'chart-palette-order-12' },
+    { value: 99, color: 'chart-palette-order-6', baseBgColor: 'chart-palette-order-13' },
+    { value: 99.6, color: 'chart-palette-order-7', baseBgColor: 'chart-palette-order-14' },
+    { value: 100, color: undefined, baseBgColor: undefined },
+    { value: 120, color: undefined, baseBgColor: undefined },
 
-      await page.setContent(htmlContent);
-
-      await expect(page).toHaveScreenshot();
-    });
-    test('render Line.Segments', async ({ page }) => {
-      const standPath = 'stories/components/mini-chart/tests/examples/score-line-segments.tsx';
-      const htmlContent = await e2eStandToHtml(standPath, 'en');
-
-      await page.setContent(htmlContent);
-
-      const segmentContainers = page.locator('[data-ui-name="Flex"]');
-      const firstSegmentContainer = segmentContainers.first();
-      const lastSegmentContainer = segmentContainers.last();
-
-      const firstContainerSegments = await (firstSegmentContainer.locator('[data-ui-name="ScoreLine.Segment"]')).count();
-      const lastContainerSegments = await (lastSegmentContainer.locator('[data-ui-name="ScoreLine.Segment"]')).count();
-
-      expect(firstContainerSegments, 'There should be only 2 renderable segments').toEqual(2);
-      expect(lastContainerSegments, 'There should be only 1 renderable segment').toEqual(1);
-
-      await expect(page).toHaveScreenshot();
-    });
-    test('render Donuts', async ({ page }) => {
+  ];
+  ScoreDonut.forEach((item) => {
+    test(`Verify with value=${item.value}  color=${item.color} baseBgColor=${item.baseBgColor}`, async ({ page }) => {
       const standPath = 'stories/components/mini-chart/tests/examples/score-donuts.tsx';
-      const htmlContent = await e2eStandToHtml(standPath, 'en');
+      const htmlContent = await e2eStandToHtml(standPath, 'en', item);
 
       await page.setContent(htmlContent);
 
-      await expect(page).toHaveScreenshot();
-    });
-    test('render correct 0 value', async ({ page }) => {
-      const standPath = 'stories/components/mini-chart/tests/examples/score-donuts.tsx';
-      const htmlContent = await e2eStandToHtml(standPath, 'en');
+      const donut = page.locator('[data-ui-name="MiniChart.ScoreDonut"]').first();
+      const semiDonut = page.locator('[data-ui-name="MiniChart.ScoreSemiDonut"]').first();
 
-      await page.setContent(htmlContent);
-
-      const firstContainer = page.locator('[id="0-1"]');
-      const donut = firstContainer.locator('[data-ui-name="MiniChart.ScoreDonut"]').first();
-      const semiDonut = firstContainer.locator('[data-ui-name="MiniChart.ScoreSemiDonut"]').first();
-
-      expect(donut).toHaveAttribute('value', '0');
-      expect(semiDonut).toHaveAttribute('value', '0');
+      expect(donut).toHaveAttribute('value', String(item.value));
+      expect(semiDonut).toHaveAttribute('value', String(item.value));
 
       const donutCircles = await donut.locator('circle').all();
       const semiDonutCircles = await semiDonut.locator('circle').all();
 
-      expect(donutCircles.length).toEqual(2);
-      expect(semiDonutCircles.length).toEqual(2);
+      if (item.value === 0) {
+        expect(donutCircles.length).toEqual(2);
+        expect(semiDonutCircles.length).toEqual(2);
 
-      for (const donutCircle of donutCircles) {
-        expect(donutCircle).toHaveAttribute('stroke-dashoffset', '0');
-      }
+        for (const donutCircle of donutCircles) {
+          expect(donutCircle).toHaveAttribute('stroke-dashoffset', '0');
+        }
 
-      for (const semiDonutCircle of semiDonutCircles) {
-        expect(semiDonutCircle).toHaveAttribute('stroke-dashoffset', '0');
+        for (const semiDonutCircle of semiDonutCircles) {
+          expect(semiDonutCircle).toHaveAttribute('stroke-dashoffset', '0');
+        }
+      } else if (item.value > 0 && item.value < 100) {
+        expect(donutCircles.length).toEqual(3);
+        expect(semiDonutCircles.length).toEqual(2);
+
+        for (const donutCircle of donutCircles) {
+          expect(await donutCircle.getAttribute('stroke-dashoffset')).not.toBeNull();
+        }
+
+        for (const semiDonutCircle of semiDonutCircles) {
+          expect(await semiDonutCircle.getAttribute('stroke-dashoffset')).not.toBeNull();
+        }
+      } else if (item.value === 100) {
+        expect(donutCircles.length).toEqual(2);
+        expect(semiDonutCircles.length).toEqual(2);
+
+        for (const donutCircle of donutCircles) {
+          expect(await donutCircle.getAttribute('stroke-dashoffset')).not.toBeNull();
+        }
+
+        for (const semiDonutCircle of semiDonutCircles) {
+          expect(await semiDonutCircle.getAttribute('stroke-dashoffset')).not.toBeNull();
+        }
       }
 
       await expect(page).toHaveScreenshot();
     });
   });
-  test.describe('Trend', () => {
-    test('render Lines', async ({ page }) => {
-      const standPath = 'stories/components/mini-chart/tests/examples/trend-lines.tsx';
-      const htmlContent = await e2eStandToHtml(standPath, 'en');
+
+  const LoadingScoreDonut = [
+    { value: 33, loading: true, color: 'chart-palette-order-4', baseBgColor: undefined },
+  ];
+  LoadingScoreDonut.forEach((item) => {
+    test(`Verify loading=${item.loading} with value=${item.value} color=${item.color} baseBgColor=${item.baseBgColor}`, async ({ page }) => {
+      const standPath = 'stories/components/mini-chart/tests/examples/score-donuts.tsx';
+      const htmlContent = await e2eStandToHtml(standPath, 'en', item);
 
       await page.setContent(htmlContent);
 
       await expect(page).toHaveScreenshot();
     });
-    test('render Bars', async ({ page }) => {
+  });
+});
+
+test.describe('Visual-Score Lines', () => {
+  const ScoreLineNoSegments = [
+    { value: 0, color: undefined, baseBgColor: undefined },
+    { value: 0.5, color: 'chart-palette-order-2', baseBgColor: 'chart-palette-order-16', animate: false },
+    { value: 1, color: 'chart-palette-order-3', baseBgColor: 'chart-palette-order-15', animate: false },
+    { value: 33, color: 'chart-palette-order-4', baseBgColor: undefined, animate: false },
+    { value: 69.5, color: undefined, baseBgColor: 'chart-palette-order-5', animate: false },
+    { value: 99, color: 'chart-palette-order-6', baseBgColor: 'chart-palette-order-14', animate: false },
+    { value: 99.6, color: 'chart-palette-order-7', baseBgColor: 'chart-palette-order-13', animate: false },
+    { value: 100, color: undefined, baseBgColor: undefined, animate: false },
+    { value: 120, color: undefined, baseBgColor: undefined, animate: false },
+
+  ];
+  ScoreLineNoSegments.forEach((item) => {
+    test(`Verify without segments with value=${item.value} color=${item.color} baseBgColor=${item.baseBgColor}`, async ({ page }) => {
+      const standPath = 'stories/components/mini-chart/tests/examples/score-lines.tsx';
+      const htmlContent = await e2eStandToHtml(standPath, 'en', item);
+
+      await page.setContent(htmlContent);
+
+      const scoreLine = page.locator('[data-ui-name="MiniChart.ScoreLine"]');
+      const valueBox = scoreLine.locator('[data-ui-name="Box"]').first();
+      const widthStyle = await valueBox.evaluate((el) => el.style.width);
+
+      expect(widthStyle).toBe(`${item.value}%`);
+
+      await expect(page).toHaveScreenshot();
+    });
+  });
+
+  const ScoreLineNoSegmentsLoading = [
+    { value: 33, loading: true, color: 'chart-palette-order-4', baseBgColor: undefined },
+    { value: 100, loading: true, color: undefined, baseBgColor: undefined },
+  ];
+  ScoreLineNoSegmentsLoading.forEach((item) => {
+    test(`Verify loading=${item.loading} without segments with value=${item.value}  color=${item.color} baseBgColor=${item.baseBgColor}`, async ({ page }) => {
+      const standPath = 'stories/components/mini-chart/tests/examples/score-lines.tsx';
+      const htmlContent = await e2eStandToHtml(standPath, 'en', item);
+
+      await page.setContent(htmlContent);
+
+      await expect(page).toHaveScreenshot();
+    });
+  });
+
+  const ScoreLineWithSegments = [
+    { value: 0, segments: 5, color: undefined, baseBgColor: undefined, animate: false },
+    { value: 0.5, segments: 3, color: 'chart-palette-order-2', baseBgColor: 'chart-palette-order-16', animate: false },
+    { value: 1, segments: 1, color: 'chart-palette-order-3', baseBgColor: 'chart-palette-order-15', animate: false },
+    { value: 33, segments: 10, color: 'chart-palette-order-4', baseBgColor: undefined, animate: false },
+    { value: 3, segments: 10, color: undefined, baseBgColor: 'chart-palette-order-5', animate: false },
+    { value: 49, segments: 50, color: 'chart-palette-order-6', baseBgColor: 'chart-palette-order-14', animate: false },
+
+  ];
+
+  ScoreLineWithSegments.forEach((item) => {
+    test(`Verify with ${item.segments} segments with value=${item.value} color=${item.color} baseBgColor=${item.baseBgColor}`, async ({ page }) => {
+      const standPath = 'stories/components/mini-chart/tests/examples/score-lines.tsx';
+      const htmlContent = await e2eStandToHtml(standPath, 'en', item);
+
+      await page.setContent(htmlContent);
+
+      const scoreLine = page.locator('[data-ui-name="MiniChart.ScoreLine"]');
+      expect(scoreLine).toHaveAttribute('value', String(item.value));
+
+      await expect(page).toHaveScreenshot();
+    });
+  });
+
+  const ScoreLineWithSegmentsLoading = [
+    { value: 3, segments: 10, loading: true, color: 'chart-palette-order-4', baseBgColor: undefined, animate: false },
+  ];
+
+  ScoreLineWithSegmentsLoading.forEach((item) => {
+    test(`Verify loading=${item.loading} Score Lines with ${item.segments} segments with value=${item.value}  color=${item.color} baseBgColor=${item.baseBgColor}`, async ({ page }) => {
+      const standPath = 'stories/components/mini-chart/tests/examples/score-lines.tsx';
+      const htmlContent = await e2eStandToHtml(standPath, 'en', item);
+
+      await page.setContent(htmlContent);
+
+      const scoreLine = page.locator('[data-ui-name="MiniChart.ScoreLine"]');
+      expect(scoreLine).toHaveAttribute('value', String(item.value));
+
+      await expect(page).toHaveScreenshot();
+    });
+  });
+});
+
+test.describe('Visual-Score Lines Segments', () => {
+  const ScoreLineNoSegments = [
+    { value1: 0, value2: 0, color1: 'chart-palette-order-2', color2: 'chart-palette-order-3', baseBgColor: undefined, animate: false },
+    { value1: 30, value2: 70, color1: 'chart-palette-order-2', color2: 'chart-palette-order-3', baseBgColor: undefined, animate: false },
+    { value1: 0, color1: undefined, baseBgColor: 'chart-palette-order-14', animate: false },
+    { value1: 0, value2: 100, color1: 'chart-palette-order-5', color2: 'chart-palette-order-16', baseBgColor: 'chart-palette-order-14', animate: false },
+
+  ];
+  ScoreLineNoSegments.forEach((item) => {
+    test(`Verify with value1=${item.value1} value2=${item.value2} color1=${item.color1} color2=${item.color2} baseBgColor=${item.baseBgColor}`, async ({ page }) => {
+      const standPath = 'stories/components/mini-chart/tests/examples/score-line-segments.tsx';
+      const htmlContent = await e2eStandToHtml(standPath, 'en', item);
+
+      await page.setContent(htmlContent);
+      await expect(page).toHaveScreenshot();
+    });
+  });
+
+  test('Verify Loading state', async ({ page }) => {
+    const standPath = 'stories/components/mini-chart/tests/examples/score-line-segments.tsx';
+    const htmlContent = await e2eStandToHtml(standPath, 'en', { loading: true, animate: false });
+
+    await page.setContent(htmlContent);
+
+    await expect(page).toHaveScreenshot();
+  });
+});
+
+test.describe('Visual-Trend Bars', () => {
+  const TrendBars = [
+    { loading: true, animate: false },
+    { loading: false, animate: false },
+
+  ];
+  TrendBars.forEach((item) => {
+    test(`Verify loading=${item.loading} `, async ({ page }) => {
       const standPath = 'stories/components/mini-chart/tests/examples/trend-bars.tsx';
-      const htmlContent = await e2eStandToHtml(standPath, 'en');
+      const htmlContent = await e2eStandToHtml(standPath, 'en', item);
+
+      await page.setContent(htmlContent);
+
+      await expect(page).toHaveScreenshot();
+    });
+  });
+});
+
+test.describe('Visual-Trend Lines', () => {
+  const TrendBars = [
+    { lastPointRadius: undefined, lastPointColor: undefined, color: undefined, animate: false },
+    { lastPointRadius: 5, lastPointColor: undefined, color: 'chart-palette-order-14', animate: false },
+    { lastPointRadius: 0, lastPointColor: 'chart-palette-order-14', color: 'chart-palette-order-14', animate: false },
+    { lastPointRadius: undefined, lastPointColor: 'chart-palette-order-4', color: undefined, animate: false },
+    { lastPointRadius: 7, lastPointColor: 'chart-palette-order-6', color: 'chart-palette-order-9', animate: false },
+
+  ];
+  TrendBars.forEach((item) => {
+    test(`VerifylastPointRadius=${item.lastPointRadius} lastPointColor=${item.lastPointColor} color=${item.color} `, async ({ page }) => {
+      const standPath = 'stories/components/mini-chart/tests/examples/trend-lines.tsx';
+      const htmlContent = await e2eStandToHtml(standPath, 'en', item);
+
+      await page.setContent(htmlContent);
+
+      await expect(page).toHaveScreenshot();
+    });
+  });
+
+  const TrendBarsLoading = [
+    { loading: true, lastPointRadius: undefined, lastPointColor: undefined, color: undefined, animate: false },
+    { loading: true, lastPointRadius: 7, lastPointColor: 'chart-palette-order-6', color: 'chart-palette-order-9', animate: false },
+
+  ];
+  TrendBarsLoading.forEach((item) => {
+    test(`Verify loading with lastPointRadius=${item.lastPointRadius} lastPointColor=${item.lastPointColor} color=${item.color} `, async ({ page }) => {
+      const standPath = 'stories/components/mini-chart/tests/examples/trend-lines.tsx';
+      const htmlContent = await e2eStandToHtml(standPath, 'en', item);
 
       await page.setContent(htmlContent);