+++ stories/components/card/tests/Card.stories.tsx
@@ -3,6 +3,8 @@ import type { Meta, StoryObj } from '@storybook/react-vite';
 
 import CardLayoutForSortableTablesExample from './examples/card_layout_for_tables';
 import DifferentCardsExample from './examples/different-cards';
+import type { TableInCardProps } from './examples/table-with-accordions-in-card';
+import TableWithAccordionsInCardExample, { tableInCardDefaultProps } from './examples/table-with-accordions-in-card';
 
 const meta: Meta<typeof Card> = {
   title: 'Components/Card/Tests',
@@ -18,4 +20,31 @@ export const CardLayoutForSortableTables: Story = {
 
 export const DifferentCards: Story = {
   render: DifferentCardsExample,
+};
+
+export const TableWithAccordionsInCard: StoryObj<TableInCardProps> = {
+  render: TableWithAccordionsInCardExample,
+  args: {
+    ...tableInCardDefaultProps,
+  },
+  argTypes: {
+    use: {
+      control: {
+        type: 'select',
+      },
+      options: ['primary', 'secondary', undefined],
+    },
+    variant: {
+      control: {
+        type: 'select',
+      },
+      options: ['card', 'default'],
+    },
+    compact: {
+      control: {
+        type: 'select',
+      },
+      options: [true, undefined],
+    },
+  },
 };