+++ semcore/date-picker/CHANGELOG.md
@@ -2,6 +2,18 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
+## [16.1.9] - 2025-09-17
+
+### Changed
+
+- Version patch update due to children dependencies update (`@semcore/base-trigger` [16.4.0 ~> 16.4.1], `@semcore/button` [16.0.8 ~> 16.0.9], `@semcore/checkbox` [16.1.2 ~> 16.1.3], `@semcore/base-components` [16.2.2 ~> 16.2.3]).
+
+## [16.1.8] - 2025-09-05
+
+### Changed
+
+- Version patch update due to children dependencies update (`@semcore/base-trigger` [16.3.1 ~> 16.4.0], `@semcore/button` [16.0.7 ~> 16.0.8], `@semcore/icon` [16.6.0 ~> 16.7.0], `@semcore/checkbox` [16.1.1 ~> 16.1.2], `@semcore/base-components` [16.2.1 ~> 16.2.2]).
+
 ## [16.1.7] - 2025-08-29
 
 ### Changed